Refactor visualization file structure
