+++ src/adaptive_SNN/models/__init__.py
@@ -1,11 +1,6 @@
 from adaptive_SNN.models.agent_env_system import AgentEnvSystem, SystemState
 from adaptive_SNN.models.base import NeuronModelABC
 from adaptive_SNN.models.environment import InputTrackingEnvironment
-from adaptive_SNN.models.metrics import (
-    compute_charge_ratio,
-    compute_conductance_ratio,
-    compute_CV_ISI,
-)
 from adaptive_SNN.models.networks.agent import Agent, AgentState
 from adaptive_SNN.models.networks.lif import LIFNetwork, LIFState
 from adaptive_SNN.models.networks.noisy_network import NoisyNetwork, NoisyNetworkState
@@ -25,7 +20,4 @@ __all__ = [
     "AgentEnvSystem",
     "SystemState",
     "RewardModel",
-    "compute_conductance_ratio",
-    "compute_charge_ratio",
-    "compute_CV_ISI",
 ]
