+++ src/adaptive_SNN/visualization/__init__.py
@@ -1,6 +1,7 @@
-from adaptive_SNN.visualization.frequency_analysis import plot_frequency_analysis
-from adaptive_SNN.visualization.plot_simulations import (
+from adaptive_SNN.visualization.api.plotting import (
+    plot_frequency_analysis,
     plot_learning_results,
+    plot_network_stats,
     plot_simulate_SNN_results,
 )
 
@@ -8,4 +9,5 @@ __all__ = [
     "plot_simulate_SNN_results",
     "plot_learning_results",
     "plot_frequency_analysis",
+    "plot_network_stats",
 ]
